async_addition/public-api/requirements.txt
fastapi 
uvicorn 
httpx 
kafka-python
pydantic